android: Subsample input hint checker histograms
In some scenarios the histogram Android.InputHintChecker.InputHintResult
can get recorded as often as frames are produced. We do not need as much
data, and we would like to reduce the overhead of recording the values.

Here, subsample the histogram on the client.

Unfortunately, this leads to somewhat uglier testing. The
InputHintCheckerTest#testInputHintResultHistogram does not know how the
random coin flip turns. Disable subsampling in this test in the most
minimal way: with side effects on other tests.

Bug: 326417367
Change-Id: Ib25a050a541de74c202bd3c57ef3712b04fc647e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6333415
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1430296}


CrOS-Libchrome-Original-Commit: 9396dfd999491fee05a6ea5abb8eb0170dcfb93f