Add modern version of base::Environment::GetVar() that returns optional
Add a version of GetVar() that returns an optional, instead of writing
the result to an out-parameter and returning a bool. Keep the existing
GetVar() but add a comment to mention it is deprecated.

Convert base/nix/xdg_util.cc and its unit test to use the new GetVar().

Bug: 400758498
Change-Id: Iff01ee45317c0c3dda211e1f0ab383eb3026f459
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6324470
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1430534}


CrOS-Libchrome-Original-Commit: 29d1bd26d171ea8d187ef62fb45519bad9551e08