[docs] Add a note to the StructTraits example about DataView usage.
Before this change it was somewhat easy to gloss over that the
StructTraits example requires the DataView version of the mojom type to
successfully serialize the types since it doesn't mention that in the
preamble to the example.

After this change we add a note to explicitly call this out.

Bug: N/A
Change-Id: I5784d09127968c697e0cbea0281148f4c9977389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6346439
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1431689}


CrOS-Libchrome-Original-Commit: 7008189bd721989e46cf9e73b2d6d5fc1c7dea47