Add base::test::IsSupersetOfValue matcher
Tests have a common need to specify a "subset" of an expected value, and
verify that some value computed by the SUT matches that expected subset
while ignoring any unexpected fields. Repeated adjacent calls to
`EXPECT_EQ(value.Find(...), ...)` using the same `value` are evidence of
this need; see e.g. the modified test in this CL.

It would be nice to have a utility that accepts a subset of a value, and
matches if a provided arg is a "superset" of that value. This CL adds
such a matcher. I will clean up some existing tests to use this matcher
in followups.

Change-Id: I2619170519312b6052075ddd92ee0b4109a24dd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387001
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437085}


CrOS-Libchrome-Original-Commit: ad88dd1f619b3ff93ae4feacf9781ef73fb9631e