[gpu] Fix media_texture_owner dump allowlist
Memory allocator dumps in background mode have an allowlist that only
supports hex codes as wild entries and is not a proper regex supporter.
This means only "0x?" is supported. There's media_texture_owner_?
which does not work properly because of this. This change updates
to use a hex code there as well so it follows the pattern matching.

Bug: 397720827
Change-Id: Id0dbac5623d9d4bafd527548a9f66d388802f500
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6394522
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1438409}


CrOS-Libchrome-Original-Commit: 02e0b427ea09f0b561c6e6dec4680f293894293e