Add C++ bindings for result type
This CL maps the result<T,E> type to base::expected<T,E> type in C++
client code. In this design, most of mojo still works with the
shadow type internally, and only transforms type at the boundary
between mojo runtime and client code. This allows most of the mojo
code to remain the same and only modifies the presentation of the
mojo result type.

There is still an ergonomic sharp edge where TWO method callback
aliases are created which I will fix in a follow up CL. This is
confusing because the "MethodCallback" alias is used by mojo, but
users have to know to use "MethodResultCallback" in their code for
methods with result return.

Change-Id: I207bb9f7e3eeb83bd96b4d432161e14d48f25daf
Bug: 404343228
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6374665
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1440052}


CrOS-Libchrome-Original-Commit: 917e4b2a2989570d44042cef4c410fae56a83032