Make task trace more useful for mojo::Shared{Associated,}Remote
Before this CL, crashing when invoking the reply callback would only
implicate internal Mojo machinery in the task trace. Needless to say,
this is not very useful the vast majority of the time, as Mojo itself is
not responsible for the contents of the reply callback.

After this CL, the task trace includes the creation location of the
shared remote. Ideally, this would be the location of the caller that
invoked the Mojo method, but that is not possible without major plumbing
changes.

Bug: 404190340
Change-Id: Ida791188a8732b4fd244a15955c6a9502e611dcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6422672
Reviewed-by: Fred Shih <ffred@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1441786}


CrOS-Libchrome-Original-Commit: f847e0452e09e8ad9b42696dca7622526a8ac068