base/uuid: get rid of FormatRandomDataAsV4
This is an odd function: it uses a subset of the data provided to it to
produce a V4 UUID. It also has no production users, so this change
deletes it and its ForTesting variant. The only use of the ForTesting
variant outside of the unit tests for base/uuid itself was in
SavedTabGroupTest, where it was being used to produce two distinct
UUIDs. Since the probability of two randomly-generated UUIDs colliding
in a test is astronomically remote (literally) and there are many
hundreds of other tests that already use random UUIDs each run, this
change migrates those tests to using random UUIDs as well.

Bug: None
Change-Id: Iba2c7413b7452340486cd4c2c11f11f57958ca76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6442401
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444920}


CrOS-Libchrome-Original-Commit: e0faa4c7861bcb675b1fc0880814cf41b42f68a7