DumpWithoutCrashing instead of DCHECK in base/metrics/sample_vector.cc
This CL changes a DCHECK to a DumpWithoutCrashing in the event that a
histogram cannot mount a sample vector reference when reading persistent
histogram data.

This is not expected (it indicates data corruption) but it's not fatal,
as the sample vector is left in a well defined state and subsequent
operations on it are also well defined.

Bug: 403816128
Change-Id: Idea7f2681b8ad6b1ca2c6cf9a85503592bda05fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6443796
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444983}


CrOS-Libchrome-Original-Commit: 81d6769660dfb2756de5946f27b0dce1b76332db