[base] Optimize base::FilePath::IsParent.
The "fast" implementation does one traversal of the 2 paths to compare
the nth component of `this` against the nth component of `child`,
until it can conclude whether `this` is a parent of `child`. There
is no first pass to extract the component of each path. There are
no heap allocations or string copies.

We keep the "slow" implementation and decide which implementation is
used with a base::Feature, to ensure that we can react quickly if
issues are found with the "fast" implementation.

Fuzzing was used to increase our confidence that the "fast"
implementation behaves the same as the "old" implementation.

In a micro-benchmark, the new implementation takes 97% less time
to execute than the old implementation.

Bug: 395595606
Change-Id: I8fc1eab690a624767d00c53ce6fa343e023f2579
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6431281
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1446368}


CrOS-Libchrome-Original-Commit: 74932f812c6d8ea154bea983010c1a7ffc1b79fe