DCHECK that histogram names are not empty.
This CL adds a DCHECK to validate that histograms are not created with
empty names. It also fixes several unit tests that inadvertently fail to
provide histogram names.

Two InstallState enum values are resurrected (kFetchSshKeys and
kMountContainer) in _DEPRECATED form to avoid leaving a gap in switch
statements that rely on the relative order, not just the absolute value,
of the install state.

Bug: 409611862
Change-Id: If1e4a4fc58d7557cc0f6574191dc0d0daa991cb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6442923
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1446581}


CrOS-Libchrome-Original-Commit: 50b79432b1cf8faf16b2f8da97a7bbdea27fbf28