[base] Populate crash keys in TaskQueueImpl::RemoveCancelledTasks.
Today, TaskQueueImpl::RemoveCancelledTasks is only called by
blink::ScriptedIdleTaskController, when it knows that there are at
least 1000 canceled tasks in the queue (in other words, it's called
infrequently). Unfortunately, a check in
blink::ScriptedIdleTaskController reveals that there are sometimes less
tasks than expected which get cleaned up. This CL adds temporary crash
keys in TaskQueueImpl::RemoveCancelledTasks to verify whether
blink::ScriptedIdleTaskController's accounting of the number of tasks
in the idle queue is accurate (depending on the answer, we'll focus
on finding issues with the accounting or finding issues in the
clean up logic).

Bug: 394266102
Change-Id: Ic9343887feead39b1feb839a6e6db0330024d4a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6357373
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1447129}


CrOS-Libchrome-Original-Commit: ca96936f65fea93f992ca9b9e3242fb0198382e2