[Android] Bind Conditions, make ConditionWithResult#get() assert phase
Make Conditions bound to either a ConditionalState (for enter/exit
Conditions) or to a Transition (for transition conditions).

This prevents incorrectly using the same Condition in multiple contexts.

Make get() assert that the Condition is bound to an ACTIVE or
TRANSITIONING_FROM state (or to a Transition) and that the result is
not null.

This makes boilerplate getters in ConditionalStates unnecessary if the field is public, but does not make the field public or remove
existing getters yet.

Bug: 394358099
Change-Id: Ibd908d36701a565f8fc958c356bbe7636e57c52d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6454450
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1447339}


CrOS-Libchrome-Original-Commit: a0644fe13fe3cf1b1223a2bbccc2f884d8b91e61