Add typemap for unguessable token
Convert unguessable token to a 32-char string. Note that the tokens
must be 32-char strings and cannot be truncated because we want to
support the following property:

  str(token) === str(mojo(str(token)))

That is, the string representation should remain stable if it is
converted to the mojo type then back.

Change-Id: Ia353910a2e7012157019569a60f0415b9df7f3cd
Bug: 406497233
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6442801
Reviewed-by: Brando Socarras <brandosocarras@google.com>
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Reviewed-by: Ashley Prasad <ashleydp@google.com>
Cr-Commit-Position: refs/heads/main@{#1448077}


CrOS-Libchrome-Original-Commit: 20eb2b846dfbe562bdab3fef7229c5439c808b70