Use base::cstring_view in base::Environment
Before this change, it was possible to pass a string that didn't
have a terminating null character. To account for this, some of the
functions created a temporary std::string, nullifying the point of
the string_view. In addition, this temporary std::string was not
created in every function.

This is now fixed by using base::cstring_view.

This change required modifying a couple existing constants to be
fully defined at the call site.

Change-Id: I5e28a4df4623d7ce9a13cfcf4b54a0194daf89d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6418733
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Amelie Schneider <amelies@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1448397}


CrOS-Libchrome-Original-Commit: 197604a7e5581884f8d017a1eca88a5070bac89d