Optimize base::DictValue::Merge
1. Skip all lookups when the dictionary being merged into is empty, and
instead replace this with the dictionary being merged from.

2. Reuse keys and value wrappers. By iterating over the dictionary
storage directly, we can move the std::unique_ptr<Value> between the two
dictionaries, rather than unwrapping it in the public iterator and
re-wrapping it in the call to Set. Similarly, this allows us to move the
std::string keys between the two, avoiding unnecessary string copies.

3. Unconditionally perform a single key lookup using try_emplace instead
of performing two in the case that the to-be-merged value is a
dictionary but the existing one is not.

Note that this is still not completely optimal, as existing entries may
need to be shifted repeatedly in the underlying vector storage (consider
`[d, e, f].merge([a, b, c])`), though this issue is not something
prominently advertised in DictValue's documentation even though it also
applies to repeated calls to Set in general. Addressing this for Merge
would drastically increase the code complexity, so we leave it for now.

Change-Id: I7ed579d846ea204b6881d8b12543f03f6e195782
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6483669
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451420}


CrOS-Libchrome-Original-Commit: 82cb8b6ee8f44d236edbd71efd9daee1fff18759