Add a synthetic field trial for using QPC on previously-uneligible devices
This will allow measuring overhead & performance impact, to evaluate if
keeping the low-resolution codepath is still relevant.

This CL also adds a command line switch, `force-high-res-timeticks`,
which tells a process whether to use QPC for TimeTicks even in the
absence of an invariant TSC. This is used by the browser process to
ensure that all processes use the same source for their TimeTicks
clocks.

Bug: 410560675
Change-Id: I4eac8a4822a1c4a689b2aac2bc6fc76db1dbdfa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6462467
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1452900}


CrOS-Libchrome-Original-Commit: 0342746e82acea0da2d4b3bc84b1525ba3a5c1c6