Use base::cstring_view in base::Environment
Before this change, it was possible to pass a string that didn't
have a terminating null character. To account for this, some of the
functions created a temporary std::string, nullifying the point of
the string_view. In addition, this temporary std::string was not
created in every function.

This is now fixed by using base::cstring_view.

Note: This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/6418733

Bug: 412679386
Change-Id: I848540b4966f8d709c39035c9d2b651aeda7f9de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6494292
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Owners-Override: Kyle Charbonneau <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1453246}


CrOS-Libchrome-Original-Commit: a29e1498627b361c1f715ddc2847145439c5e3f9