[Android] Make ElementFactories depend on Elements and not Conditions
ElementFactories only make sense when they require something that isn't
produced yet. They could depend on Condition, but there is no practical
use for this - it always makes sense to wait on a ConditionWithResult,
and it's simpler to always depend on an Element.

Elements' Conditions will be able to be set soon, and that would cause
ElementFactories to wait on Conditions that won't be part of the
transition anymore.

Bug: 414438679
Change-Id: I1dd5b7bfe394b3ae27bd77855e848d591aad960a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6500920
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Owners-Override: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454556}


CrOS-Libchrome-Original-Commit: 6cb04b443544d604c4c09fbfc2baa126edf21ff8