Set symbol_level=0 on asan and tsan linux CQ bots temporarily
Current infra failures are due to the test bots not being able to
download the cas bundles, eg:
https://ci.chromium.org/ui/p/chromium/builders/try/linux_chromium_asan_rel_ng/2339427/infra

It could be related to file size, so this tries shrinking the test
binaries on these two bots a little while we figure out why siso/cas
is misbehaving.

This also disables ToolsSanityTest.DataRace since it was implicitly
depending on there being symbols in the resultant stack trace
(specifically the test filename). So it fails when symbols are
removed.

It's unfortunate to lose symbols in the test failures on these bots,
but we presume that's better than a) remove the builders from the CQ
all together or b) leave them flakily infra-failing on random CLs, eg:
https://ci.chromium.org/ui/p/chromium/builders/try/linux_chromium_tsan_rel_ng/2329490/infra
https://ci.chromium.org/ui/p/chromium/builders/try/linux_chromium_tsan_rel_ng/2329515/infra
https://ci.chromium.org/ui/p/chromium/builders/try/linux_chromium_tsan_rel_ng/2329502/infra

Bug: 416191043
Change-Id: I744f962f57878e07731c45201eec51a4c260f648
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6520153
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1457260}


CrOS-Libchrome-Original-Commit: 6b6e6e8a97cd97e9bca38c4a7f5ba61066002967