Add SendValidation tag/attribute to Mojo
This change adds a new mojo tag/attribute to make it easier to debug
validation errors on serialization in the case where there's developer
error. This takes advantage of the mojo Feature infrastructure to flip
the validation error on when the feature is on.

```
The tag can be added like so:
feature kBarValidationFeature {
  const string name = "BarValidationFeature";
  const bool default_state = true;
};

interface Foo {
  [SendValidation=kBarValidationFeature]
  Bar(Struct struct_needing_validation);
};
```

This change adds unit tests in mojo_unittests.

As a part of this change, another set of serialization templates were
added to facilitate a smaller binary size. If SendValidation is
requested, the alternative template is used. If no SendValidation
specialization exists, then the code will automatically fall back to the
original serialization.

The compile size is greatly expanded because this adds a
new serialization path to every Struct, Union, Array, and Map used by
Mojo. A follow-up change reduces increase down to about 0.45% compile
size increase with additional complexity.

Compile-size: See commit description.
Bug: 400895804
Change-Id: Idb0c4daeb179e938ea4c9e195f6c95f0e45c8f04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6328899
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Emily Andrews <emiled@microsoft.com>
Reviewed-by: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1457638}


CrOS-Libchrome-Original-Commit: 92983e93b50fbb20af78d5339d03d727eee439c5