Add separate generated <module>-send-validation.h file
This change adds a separately generated <module>-send-validation.h file
which will be included when a target uses send validation. The file must
always be generated to ensure the build outputs are deterministic, but
when there is no send-validation, it can be excluded as an include.

To facilitate testing the Serialization code path, an
IncludeSendValidation module attribute was added. This causes all the
generated send-validation.h required to serialize with send validation
in the <module>-shared.h file.

This change also required splitting out *DataView declarations into a
different place so they could be shared in the
<module>-send-validation.h file and in the <module>-shared.h file.

Increase in compile size is significantly reduced from previous change.
This may be unavoidable because of the SFINAE.

compile-size: see message.
Bug: 400895804
Change-Id: I487a3e52c88996e797f661055b3f870f433f7cad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6489356
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Commit-Queue: Emily Andrews <emiled@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1457639}


CrOS-Libchrome-Original-Commit: a543dd4ebed1c7abc66739418691b8392fc9cca7