Add handle type allowlist for transfer to untrusted process
This CL adds a new allowlist for types of handles that can be
transferred to an untrusted process.

The allowlist has been determined empirically and is:

Section
File
Directory
DxgkSharedResource

A default-enabled feature is emplaced to control this new protection
called MojoHandleTypeProtections, in case of any issues.

A test is added to verify that a NOTREACHED occurs when attempting to
transfer this type of handle.

BUG=405143032

Change-Id: I6f0fa6887beb1766cf19fdf0d19ee3487dd160a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6379023
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1457709}


CrOS-Libchrome-Original-Commit: 3cb6f6c2e3de7eb17fdde9d3bd4637038f5de479