[MappableSI] Fix V4L2VideoEncodeAccelerator::SetCommandBufferHelperCB().
Fix V4L2VideoEncodeAccelerator::SetCommandBufferHelperCB to use a
blocking call which is posted on |encoder_task_runner_|. This is
required for SharedImageInterface to be initialized before its being
used.

Note that blocking on |encoder_task_runner_| since it is not IO thread
or media mojo thread.

Bug: 393315127
Change-Id: I79b2e48f1b7438a5c27abd92f160aec5bc031ab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6544486
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: vikas soni <vikassoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460190}


CrOS-Libchrome-Original-Commit: 0cc39c97b63547119a5b6b570963c5d8b46773de