Turn on java goldens for mojo
This also changes the behaviour of the mojo java generator. The
--java_output_directory will now skip generating the *srcjar file if
the flag is provided. It seemed kind of weird that it would produce
both the srcjar and the src file if the argument was provided.

I could not find any usages of this flag anywhere in the codebase.

Change-Id: Ie09e423956be68b709c6202c0f25effbb4c684a2
Bug: 40841428
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6547999
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1464410}


CrOS-Libchrome-Original-Commit: 801ccaa41f2ac5541e50f2fb6c37e62a9dc2941d