Fetch user remote command only when CEC is enabled
For now, we rely on if DM server has set user remote command
topic in Policy data.

However, this solution is hacky as the topic field is not
designed for this purpose and will be deprecated in the future.

Long term, we should make DM server to use `cec_enabled` which is a more
specific signal that indicates the CEC state of the domain.

Bug: 419832738
Change-Id: I3c67e3c5635f38463d0725699b42c4c72e86773d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6579328
Commit-Queue: Owen Min <zmin@chromium.org>
Reviewed-by: Yann Dago <ydago@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1465026}


CrOS-Libchrome-Original-Commit: 053240cd8a9e7c4465aebcb899329d56dc53c2a6