GWP-ASan: raise kMaxMetadata and kMaxRequestedSlots
kMaxMetadata used to be used as the size of a statically allocated
array, but it isn't anymore, and there's not much good reason for it
to constrain the maximum number of metadata slots kept around.

There's some potential, especially on Android, for the GWP-ASan
region to fail to be allocated as the PA pool may be too fragmented.
Region allocation failure use to be crash, now it is handled gracefully
and this is taken into account by the Security.GwpAsan.Activated.*
histograms.

This sets kMaxReservedSlots as approximately the 16-bit max,
kMaxRequestedSlots to half that (as required), and kMaxMetadata
to half of that, since some tests rely on kMaxMetadata being less
than kMaxRequestedSlots. In the future, the tests should be changed
to use test-specific values instead of the max values, and then
nothing needs to constrain kMaxReservedSlots and kMaxMetadata except
the limits of a 16-bit integer.

Bug: 417503031
Change-Id: I84346e61dc3e61e36b19be33e625574829109f8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6545318
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Sergei Glazunov <glazunov@google.com>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1466134}


CrOS-Libchrome-Original-Commit: 7e9bbf0ed82a70ec80f3816c51578cad4e4bdaad