Automatically join threads in hang_watcher_unittest.cc
Test in hang_watcher_unittest.cc were manually managing the lifetime of
the spawned thread. It's safer and simpler to leverage RAII and simply
close the threads when they go out of scope.

If the thread is joined in ~BlockedThread, tests can't also manually
Join the thread in the test body or else the thead would be incorrectly
joined twice. Instead, tests can call `WaitDone()`, which waits until
the thread terminates, but without actually joining it.

Bug: 413314866
Change-Id: I7dcf744c8c74a5a09f2b8b8228c3ffcc956aa684
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6583574
Reviewed-by: Olivier Li <olivierli@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1466473}


CrOS-Libchrome-Original-Commit: bb7bb9e5e7a143a4a28087c1c6b528389ca5a158