Remove superfluous includes for base/debug/stack_trace.h
Code in base/debug/stack_trace.h is in namespace base::debug, so the
lack of "debug::" and "stack" in a file are strong indicators that a
file does not need this include. Remove it and then do IWYU to fix the
build.

Bug: 40318405
Change-Id: I7143f3fbea9a6653cbb5935e9ea2549ee3135a93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6622838
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Owners-Override: Kyle Charbonneau <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1470133}


CrOS-Libchrome-Original-Commit: 96f6f27c5c80394e43a7bd86a57338b6c3005b0c