Add a LockFreeAddressHashSet mode that stores keys in blocks for better locality
This meets all the guarantees of the original LockFreeHashSet because
all modifications and the Copy operation are still done under a lock,
and the access patterns of Node objects hasn't changed:

* New nodes are always added at the front of a chain, and never deleted,
  so reader threads with pointers to the original node can still walk
  the chain starting from that node.
* The only modifications to the values in nodes are atomically replacing
  NULL with a key value (to insert a key), and atomically replacing a
  key value with NULL (to delete a key).

In multi-key mode, a reader thread with a pointer to a node iterates
over ALL key slots in the node instead of reading a single slot. Since
the reader doesn't make any control flow decisions based on the slot
contents, each atomic read has the same possible outcomes as the atomic
reads of a single key in single-key mode.

Also keeps the original one-key-per-node mode for A/B testing.

Bug: 383374205
Change-Id: I8f997a7ccf7a062a26b72c131028db2cdcf0297d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6197096
Commit-Queue: Joe Mason <joenotcharles@google.com>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1472458}


CrOS-Libchrome-Original-Commit: 24260588f1966e92756d7b13437902d93384a0c0