BFCache: Add kSharedWorkerMessage eviction reason
This CL adds `kSharedWorkerMessage` reason, a new BFCache eviction
reason. This is triggered when a BFCached page receives a message from a
SharedWorker. The eviction logic will be implemented in a follow-up CL
so that this change can focus solely on adding the new reason.

For more details:
https://docs.google.com/document/d/1CI7iqg_NgK7NELOzK6jrF_Zk-FATilMwuO53M4ZApVg/edit?usp=sharing

Bug: 406420935
Change-Id: Iedd247020d7663c82243e5abcf51130ab4292643
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6640810
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Anna Sato <annasato@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1477780}


CrOS-Libchrome-Original-Commit: 21f0fee9f19e54d3e958086ce3164b691a0cf6c2