Remove dependency of flat_tree on span
This implies that plain arrays cannot longer be passed directly into
a flat_tree. But, if necessary, users can use a span or an initializer
list.

The goal of this CL is to reduce the compile size contribution of
flat_tree.h. Span accounted for almost 90% of it.

The compile-size bot reports an improvement of 36MiB.

Bug: 40284755, 426576237
Change-Id: Ife52d35febda1823f71a13d91f32a05d854d9637
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6653082
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Aldo Culquicondor <acondor@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1478139}


CrOS-Libchrome-Original-Commit: db03e1f986c68609c57e3713c5de6d54af524e6f