Correctly mark ThreadGroupProfiler blocking
On Android ARM32 one of the unwinders creates a memory-mapped file,
which is a blocking operation. We're already on a separate task runner,
so we just need to annotate that this may block. This fixes the
android-rust-arm32-rel bot.

Bug: 427468122
Cq-Include-Trybots: chromium/try:android-rust-arm32-rel
Change-Id: Id7738597a0e13b04e046772507a02a7824368f16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6691659
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480780}


CrOS-Libchrome-Original-Commit: 584502c158521aeafe60d6bad641f8ea80b2d7db