Path service: add method to override path with a maybe-invalid value
It's possible for some of the path to be relative or even invalid in
real life, however, those cases were not covered in test because path
service override ensure the validity of the path set.

This CL adds methods only for testing to override the path with a value
that might be invalid.

See the associated bug for more information.

Bug: 428455312
Change-Id: I2de1b0515ef68d9d81ddb8618075c99a447b8350
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6698535
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Steve Becker <stevebe@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1482113}


CrOS-Libchrome-Original-Commit: 814f95a984b8f07376e686159d8c206f838ffb3c