install-build-deps: Use `logging`
The real purpose of this CL is to make the wall of text emitted at
line 877 (error condition in `install_packages()`) more readable.
Decode it to a regular `str` so that the newlines are treated
properly.

Additionally, join the list of required packages in
`find_missing_packages()` with newlines to form a single list.

As for logging changes, rather than repeated invocations to
`print(..., file=sys.stderr)`, instantiate an explicit `logger` and
configure it to spew on stderr. Convert all `print()` statements
accordingly.

Change-Id: I195ee1e110aff423c59cb476915724ad97caba26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6702870
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1482258}


CrOS-Libchrome-Original-Commit: aedef3ab338bd43b0fa2f9b51096749f3bfc2815