Revert "trace_event: trace_arguments: fix risky memcpy"
This reverts commit fa9611a35db47e240096c384c375da92890cb5fc, which was
submitted here https://crrev.com/c/6632927.

Reason for revert: This change may not be submittable to upstream, so it
is being converted to a long-term patch for the time being.

If, however, this change is submitted to upstream, then in a future
libchrome uprev that contains this change, the long term patch will be
detected and deleted.

The patch is located at:
libchrome_tools/patches/long-term-3000-trace_event-trace_arguments-fix-risky-memcpy.patch

BUG=b:387993322
TEST=cros-workon-asurada start libchrome; emerge libvda-gpu-tests.

Change-Id: Ic7c474b6cfb90973b5980615781a7081e36a3ea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/libchrome/+/6702838
Reviewed-by: Georg Neis <neis@chromium.org>
Tested-by: Di Wu <diwux@google.com>
Reviewed-by: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Commit-Queue: Di Wu <diwux@google.com>