[base] Introduce base::ByteCount.
Existing code often manipulates byte values (KiB, MiB, GiB) using raw
numbers, leading to:
- Ambiguous units: The intended unit is often only documented in
comments, which are prone to errors and are not enforced by the
compiler.
- Error-prone conversions: Manual conversion between units (* 1024 *
1024) is frequently duplicated, increasing the risk of bugs and making
code harder to maintain.
- Reduced readability: Code manipulating raw numbers for bytes is
difficult to read and understand.

We should introduce a base::Bytes class, similar to base::TimeDelta, to
encapsulate byte quantities.

Example situations that will be improved with this API:

https://source.chromium.org/chromium/chromium/src/+/main:base/win/hardware_check.cc;l=125-126;drc=3ec0d9a35012c30f63b9f3bf8fd013df67a292b8
Due to unit confusion, this code checks that RAM is >4MB, but intended
to check that it's >4GB.

https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/policy/site_isolation_policy_browsertest.cc;l=287-292;drc=83bd66b5413315522137d61ff3f2a85bad916dc5
This function takes a "ram_kb" argument and passes it to the constructor
of base::test::ScopedAmountOfPhysicalMemoryOverride, which expects an
amount of RAM in KiB.

Bug: 429140103
Change-Id: Ib816c6ca2fd305a8940fe2bcf215801ac4dc957c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6689072
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1482705}


CrOS-Libchrome-Original-Commit: 714df63ec8dd8de12317d533912ac4424a227b91