Update HangWatcher to support the GPU process
RunOtherNamedProcessTypeMain and RunZygote in content_main_runner_impl.cc are
already starting a HangWatcher for the GPU process if it is unsandboxed. This
had no effect however because HangWatcher::InitializeOnMainThread was aborting
for the GPU process. With this CL, HangWatcher will now run for unsandboxed GPU
processes if the feature "EnableHangWatcherOnGpuProcess" is enabled.

A follow-up CL will enable HangWatcher for sandboxed GPU processes.

Change-Id: I02b24046bed46cd2a32f53c7389700eb5134b001
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6640490
Reviewed-by: Olivier Li <olivierli@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486296}


CrOS-Libchrome-Original-Commit: 11f6c808c46b95084a84ffb665384d3ed942dc7a