[mojo] Refactor bindings generator
Change the processor to be immutable and hoist some control flow out of
the processor itself. This makes calls to the processor idempotent which
makes it a lot easier to reason about.

Also refactor the typemap verification to work outside of the processor.
This decouples the verification logic from the generation logic and
allows both components to work separately from each other.

Change-Id: I04ee8c1176b456ba269c653952c46d2ee3a96d31
Bug: 428039213
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6721012
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486488}


CrOS-Libchrome-Original-Commit: 403b3648843ed2529f0dbd2becc4fa485c7c9f05