Add Enterprise Policy ProtectedContentIdentifiersAllowed
ProtectedContentIdentifiers are shared to web pages in certain scenarios
to access some content. This involves sharing a distinct identifier, so
this CL defines an enterprise policy that prevents the sharing of this
distinct identifier.

When unset, protected content identifiers are allowed.

This is tested via unittests, and browser tests which verify that when
the policy is set to false, software secure key systems work as expected
but hardware secure key systems which generate a distinct identifier do
not work.

Manually tested by verifying hw secure playback works when policy is
unset or set to true, and verifying hw secure playback does not work
when policy is set to false. This was tested on Windows.

Playback also works for L3 and L1 Android without any errors.

Testing doc here: http://shortn/_Ebp5cn6hRb

Bug: 424007290
Change-Id: Id3160be88b032264a2edf0b36ac0eb60868fae4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6716602
Reviewed-by: Sangbaek Park <sangbaekpark@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Artem Sumaneev <asumaneev@google.com>
Cr-Commit-Position: refs/heads/main@{#1487081}


CrOS-Libchrome-Original-Commit: 6b07e7b77ae41a7b2946f2d02d314135a8e00c9d