Remove failing check in UpdateRecursiveWatchesForPath()
In the inotify base::FilePathWatcher implementation,
UpdateRecursiveWatchesForPath() has a DUMP_WILL_BE_CHECK() that is
generating non-fatal crash dumps. This shows the invariant does not
hold, but the lack of follow-on issues shows the code otherwise works.
As such, just remove the failing DUMP_WILL_BE_CHECK() and add a note to
explain what is going on.

Bug: 427038829
Change-Id: I1753fa3f1001ad05f13c5f1abc2a49ec5cdc17a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6762452
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Momoko Hattori <momohatt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1488334}


CrOS-Libchrome-Original-Commit: 2afddb26c8aea583d9fb1e226372daf5d1f43dbd