Add javaless renderers behind GN arg
This CL adds support for javaless renderers, a feature that is coming to
Android in the near-ish future (months, not years). It is hidden behind
the already existing `enable_javaless_renderers` GN arg.

The bulk of this change exists in javaless_child_process_service.cc.
Many of the other small changes are directly related to usages in
javaless_child_process_service.cc.

We also add a second <service> entry in the manifest (when this GN arg
is true), which will likely need to be tweaked as Android updates their
implementation.

Since Android does not have this officially released, there is no
native_service.h in the NDK. Instead, I've copied the current prototype
version of what may end up in the NDK directly into
//third_party/android_toolchain. We will remove this once the NDK
includes native_service.h.

Currently, with the latest prototype available to me, this shows
websites with enable_javaless_renderers=true.

Bug: 427481060
Change-Id: Ib6f597734423d97caeadba11b7673f836933c24e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6287425
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Sam Maier <smaier@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1490863}


CrOS-Libchrome-Original-Commit: 808f006925c1f83ebc9f3cb9330d3f90d9e7c7fb