crash: include command output in minidumps to debug a net problem
When the CHECK tracked by bug https://crbug.com/40064248 trips, a magic
crash key (net-crbug_40064248) is set, which chrome_crashpad_handler
will act on by recording a special CommandOutputUserStream in the
minidump. This stream can contain the output of multiple commands. In
this case, the commands that are executed are system_profiler, ifconfig,
systemextensionsctl, and lsof.

This is intended to be temporary, and will only be used to collect
information about systems that are susceptible to the bug, to determine
whether they have an unusual configuration in common.

https://chromium-review.googlesource.com/c/6791935 implements a decoder
for the new temporary command output stream (type 0x4b6b0005).

Bug: 40064248
Change-Id: Ie6c0d921228482b6d3f27216fa949ba8a3fef9f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6788541
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Mark Mentovai <mark@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1493431}


CrOS-Libchrome-Original-Commit: f6ee10c0369d457f0ebec4d43a47063065384f50