[mojo] Add which* API for mojo unions in typescript
Previously, users would have to check for the nullness of the entire
object to figure out which field is set. This is somewhat error prone
because not everyone considers the malformed union case. The limitation
of the ts typesystem means that we cannot enforce that one and only one
field is set on an object. Now users can do:

switch (whichMyUnion(u)) {
  case MyUnionFieldTags.BLAH:
    ...
  case MyunionFieldTags.ZWAH:
    ...
}

This pattern is better because it will actively flag a switch statement
that is missing. This is hugely beneficial when detecting code that
needs to be updated when adding a new field to a union.

Unfortunately, a problem with this approach is that we cannot catch a
malformed union that has TOO MANY fields set.

The function approach was chosen because mojo ts types are interfaces,
which means that we cannot rely on any specific field in js object to
figure out which field is set.

**note** although it is possible to catch missing switch cases, we
currently do not have that ts compiler check enabled (at least not
for the test..).

Bug: 432105758
Change-Id: I85952dd16cd0350c13779ac42bb49cd91c00d526
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6799046
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1494496}


CrOS-Libchrome-Original-Commit: ec7028ead710b969da432025acbeea42a43dacdd