base: Add integer overflow checks in File::Read/Write
This CL adds integer overflow checks in the `File::Read()` and
`File::Write()` methods on POSIX system. base::CheckAdd utility is used
to validate the calculation of the final offset does not result in an
integer overflow before the value is used. Without this check, there is
a potential risk with very large offset and size values.

Bug: 391398191
Change-Id: I0bea697d226959316c4845708b4dbbde14b664e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6809865
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1495299}


CrOS-Libchrome-Original-Commit: 41905e418351fe5fa6266f14d109834ada25a78f