Fix sync mojo call hang when associated receiver is disconnected
Previously MultiplexRouter would only unblock a sync call when the
receiver peer is disconnected if the entire task queue was empty. This
would mean that a sync call can hang indefinitely if anything else
is in the queue.

This CL fixes this by processing peer associated endpoint closed
events for endpoints that are waiting for a sync call reply
immediately, even when the task queue is not empty.

Bug: 434802068, 435493653
Change-Id: Iae80b72b37d191edde73cdfbab5a4effe33f2bcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6807214
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496636}


CrOS-Libchrome-Original-Commit: 262991c522f8ffec290c6ffd616f9fa8791ec30f