libchrome: delete backward patch 2400, iterate patch 2100
Backward compatibility patch 2400 is no longer needed after the recent
Abseil upgrade. So delete it.

The current 2100 patch is deleted too so that board grunt is exposed to
the newer APIs exposed by the recent Abseil upgrade through the grunt
board's mojo -> ipcz dependency chain. However, while the newer
absl::Overload API was introduced and exposed by the recent Abseil
upgrade, the API is still not new enough to be inline with the one used
in upstream Chromium. As a result, a different 2100 patch is added in
this change to patch this slight absl::Overload difference between the
ChromeOS' current Abseil version and the Abseil used by upstream
Chromium.

To future code contributors, please note that patch 2100 needs to be
tested against board grunt as currently grunt is the only board that has
ipcz enabled. Other boards compile just fine with or without the old and
the new patch 2100.

BUG=b:425266540, b:425798059
TEST=emerge-brya libchrome

Cq-Depend: chromium:6831793
Disallow-Looks-For-Green: True
Change-Id: Ia2450e73c964439edfc6abad9bbe88ce10e1825c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/libchrome/+/6826173
Commit-Queue: Di Wu <diwux@google.com>
Tested-by: Di Wu <diwux@google.com>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>