Preparation for async MemoryPressureListener registration
In a future change, the registration of MemoryPressureListeners
will be made asynchronous, by posting a task to the main thread.

Some test suites fail when registration is asynchronous, because
they use SimulatePressureNotification, which posts the notification
tasks synchronously, without waiting for the registration of the
actual listener to complete.

This CL introduces SimulatePressureNotificationAsync, which posts
the notification tasks asynchronously on the main thread, ensuring
it executes after any queued listener registrations are complete.
Those tests already pump the message loop to handle the actual
async notifications, so no further test modifications are needed.

Additionally, NotifyMemoryPressure() is now always posted to the main
thread. This prepares for a future change where the listener observer
list will no longer be thread-safe.

Bug: 436324601
Change-Id: I2ce8365dd6b345f08da325a4082038cfd23a40cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6837378
Reviewed-by: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Victor Vasiliev <vasilvv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501376}


CrOS-Libchrome-Original-Commit: b7acffac48c5ba62e7575c4ad610b26a15fdce38