[MemoryPressureListener] Move all registrations to the main thread
Before this change, listeners were held in a TreadSafeObserverList,
which allowed them to be registered from any thread. This had the
drawback of making all notifications asynchronous, via a PostTask.

SyncMemoryPressureListener exists for the few cases where the
notification must arrive as soon as possible, but it involves holding a
lock to send the notification.

This CL moves all registrations to the main thread, by switching the
ObserverListThreadSafe with a regular ObserverList that lives on the
main thread. This means it's now possible to register a synchronous
listener without involving any lock, with the drawback that the
listener must also live on the main thread. This is fine as all current
users of SyncMemoryPressureListener already live on the UI thread.

To support listeners living off the main thread, MemoryPressureListener
was refactored so that it now internally owns a
SyncMemoryPressureListener, which is bound to the main thread, and
communication is done through posted tasks.

Bug: 436324601
Change-Id: Icd102d2c696b4af31d2d40590509bf0943655f0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6817979
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501434}


CrOS-Libchrome-Original-Commit: be7c9f294902260356b771f3104fdeff27095f01