Remove IPC::ChannelProxy::AddListenerTaskRunner().
Also do the same for RemoveListenerTaskRunner().

They are uncalled. In turn, this makes the same methods in the
ChannelProxy::Context become uncalled. In turn, this makes the listener_thread_task_runners_ list always empty, so remove it.
In turn, this makes that lock that guards it unused.

Change-Id: Ic0f112bc9cd0183f30abfb732547f39b82d266cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6853175
Auto-Submit: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1502115}


CrOS-Libchrome-Original-Commit: 54cc0aad7395ead70244aa6a920a62ccf16bb1da