[NTPShortcuts] Add NTPShortcuts policy to pref mapping
Adds business logic for mapping the NTPShortcuts policy to the
`enterprise_custom_links.policy_list` preference. The
NTPShortcutsPolicyHandler performs validation checks for policy entries
and converts the policy dictionary to a dictionary in the profile
preferences that matches the `Links` struct.

The policy handler is gated by NTPShortcutsPolicy feature which is
disabled by default. When the feature is disabled, the policy handler
will ignore the policy and set the preference as an empty list. When
enabled, the preference will be set based on the policy values.

Reuses existing field_validation_test_utils for policy handler testing.

TODOs have been added to update the hard-coded preference dictionary
keys once the new EnterpriseLink struct is implemented.

Change-Id: I5f844b81590a415ee3915ba4d4b962c1015c38ea
Bug: b:437359067
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6775107
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Alex Chen <alexwchen@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1504227}


CrOS-Libchrome-Original-Commit: 77d9c2d34f6584b16d53ba60010afc172f212e0c