[Android] Migrate some tests to PausedExecutorTestRule
These tests rely on code running on a background thread. Robolectric
tests work best when we control the execution order of async code
(like AsyncTasks), and this CL migrates them to use PausedExecutorService
as the thread pool executor to do so.

This is a preparation to switch all Robolectric tests to use
PausedExecutorService as thread pool executor. These are the tests that
failed with the upcoming change and can be fixed by manually calling on
the PausedExecutorService to run.

Bug: 40934211,41495939
Change-Id: I505435527e559f7717e14d68c69a68dce6ade273
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6867692
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Owners-Override: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1504702}


CrOS-Libchrome-Original-Commit: 2da9490076ea93d8a5849e58eff0d01931a9fac1