[Mojo] Fix final (?) sync call on associated interface hang.
When we detect that a peer is closed the corresponding sync_watcher_
is signalled. However that will only call the OnSignaled callback is
the watcher is currently being waited on. If instead a different
same-sequence watcher is being waited on, SequenceLocalSyncEventWatcher
will clear its list of ready watchers. By re-signalling `syng_watcher_`
in MultiplexRouter::Endpoint::SyncWatch if peer_closed_ is true,
while ignoring that `sync_message_event_signaled_ might already
also be true, we can make sure not to block in this edge case.

Bug: 436965298, 434802068
Change-Id: I61bc064eebfccac58ebb74879c3e23ffe019d54b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6851904
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1505136}


CrOS-Libchrome-Original-Commit: 08eaac24f17738f9de49313c38f2867a2db883d9