[Android] Use PausedExecutorService to run bg PostTasks in Robolectric
After this CL, BaseRobolectricTestRule/Runner will always install
PausedExecutorService as PostTask's
sPrenativeThreadPoolExecutorForTesting.

This is equivalent to putting PausedExecutorTestRule in all tests.
Removing redundant PausedExecutorTestRule will be done in a subsequent
CL.

It can be uninstalled (and I've done so for a few tests which are lower
level and where we don't want this).

Bug: 40934211,41495939
Change-Id: I333f8610443f53f999c71ec8614cbc529868728e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6849495
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Owners-Override: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1505280}


CrOS-Libchrome-Original-Commit: 5bc8bb06c5ce3c6c3608a0a5c76e93a682ecfe66