base: Speculative crash fix in ScanStackForNextFrame
In `ScanStackForNextFrame`, the calculation of `last_fp_to_scan`
could underflow if there is an unexpected overflow of
`fp + kMaxStackScanArea` and `std::min()` results in an overflown
low value. In this case, `last_fp_to_scan` could be unreasonably
large and leads to an out-of-bounds read and a SIGSEGV.

The CL uses `base::ClampedNumeric` to safely calculate the upper
bound for the stack scan. It also adds a few crash keys in case
the speculative fix does not work.

Bug: 402542102
Change-Id: I03d306ac7a6ea4269071db4fa21c58c1b87ff30f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6896266
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508425}


CrOS-Libchrome-Original-Commit: 6acadb43fdf5769a3be3751d1979180c1f9a48c0