ipcz: Add log messages in router and node_link
Router is a central abstraction in ipcz. It is important to know about
all these objects when investigating ipcz bugs. Lifetime of Router is
also closely related to lifetime of Sublink.

Add a few log messages to make it easier to track lifetimes of Router
and RemoteRouterLink.

Bug: None
Change-Id: If2a9a7eca7dec553480905156c083f25008b26f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6905431
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1509482}


CrOS-Libchrome-Original-Commit: 41d18ff643fd0b54a6e87c6f3864ea49a6088ae8