Add DirectReceiver attribute for mojom and annotate interfaces
Adds a [DirectReceiver] attribute for mojom interfaces. This attribute
must be provided if an implementation of an interface is bound to a
DirectReceiver<T>. This is enforced at compile time.

Additionally, any interfaces that become entangled with the direct
receiver must also be marked as [DirectReceiver], for instance if they
are provided as a pending_receiver<> to a method parameter of another
DirectReceiver. This is enforced at compile time using a mojo bindings
check.

Finally, all interfaces that need the annotation today are provided with
the [DirectReceiver] mojo attribute. As all checks are done at compile
time, this CL should not have any functional effects.

Bug: 438226500
Change-Id: I233191a854a55785cd3a60257a3472b1df521df6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6897496
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1509781}


CrOS-Libchrome-Original-Commit: daa8ba079e8cbe00a9f98b016cc727e5808f5286