win: Add dirs allowed to leak to test_file_utils
This CL adds a function to base/test_file_utils which allows test infra
to get/set an allow list of sub-directories that are allowed to fail the
DieFileDie() call in OnTestEnd. If all the non-deleted files contain one
of the allowed sub-directory paths, then the failure to delete the
directory is ignored.

InProcessBrowserTests on Windows will start with an initial allow list
of "\\Sync Data", L"\\Local Storage\\leveldb", L"\\DataSharing", and
L"\\Collaboration". This allows us to still get most of the benefit of
file leak detection without disabling a lot of tests. The plan is to fix
the issues with those directories one by one, and remove them from the
allow list. If, instead, we need to add even more directories, we should
consider removing the file leak detection, at least for Windows.

Also, reenable ManagedUiTests, the first suite of tests that were
disabled when this issue started.

Bug: 410751413
Change-Id: I744ff331f652010ad9d0cef45a58738cc7a7f2ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6794988
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1510938}


CrOS-Libchrome-Original-Commit: fec3d9db9b981fdb44cf0ce4e7dc5a997f82e58e