Reset `ScopedFeatureList` in `BlinkTestEnvironment` destructor.
This change ensures that the `ScopedFeatureList` instance managed by
`FeatureListScopedToEachTest` is reset when `BlinkTestEnvironment` is
destroyed. This is necessary because if `exit()` is called within a test
(e.g., in fuzz tests), `BlinkTestEnvironment`'s destructor can run while
`FeatureListScopedToEachTest` is still active, potentially leading to
incorrect destruction order of `ScopedFeatureList` instances and
triggering CHECKs. A static instance pointer and reset method are added
to `FeatureListScopedToEachTest` to allow this forced reset.

Note: It may seem redundant that there's two levels of
ScopedFeatureLists (one in FeatureListScopedToEachTest and one in
BlinkTestEnvironment), but per the CL desc of this CL that added it,
it's apparently necessary: https://crrev.com/c/5739000

Bug: 441543203
Change-Id: If5c13040963d5c09d570ca02af67d89853172cfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6915483
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1511540}


CrOS-Libchrome-Original-Commit: c8f6f724803cd8de66fc447f432feb50eb2aa942