Replace Java call used for determining background media suspension
This change cleans up the runtime Java call in renderer, that was used
to enable background media to continue playing for large form factors.

This also introduces a new "largeFormFactor" device info category, that is intended to capture any device with a screen >= 600dp.

Lastly, we removed the temporary guard that prevented the Java calls for webview

Bug: 438910961
Test: Manual (desktop, tablet)
Change-Id: I826186b6d66f407131df1a65449d22c00cb95353
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6898216
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Eric Lok <lokeric@google.com>
Cr-Commit-Position: refs/heads/main@{#1512796}


CrOS-Libchrome-Original-Commit: d8db38d62575135afbae163cdae12be517e8ef39