Prevent v8 symbolizing stack traces in sandboxed processes
V8_Fatal attempts to print out a stacktrace even in sandboxed
processes where this results in nothing happening. When dbghelp
is not available (i.e. in sandboxed processes), this results in
a crash during symbolization that masks the V8_Fatal error.

This CL removes the stack trace printer from the V8 platform
in processes on Windows where stack dumping is not enabled.

Bug: 443286263
Change-Id: I52ff3af37311a73b65e4343a73bcbf40bc3b7333
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6933178
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1514393}


CrOS-Libchrome-Original-Commit: d007d6283efbcbcb506ce9a37536d8c9d8a9bf9c