Refactor ChildProcessConnection to use a controller for bindings
This change introduces a new abstraction layer,
`ChildServiceConnectionController`, to manage the service binding logic
within `ChildProcessConnection`.

The primary motivation is to support a new upcoming Android API, which
allows modifying the flags of an existing service binding. This avoids
the need for the previous approach of maintaining multiple parallel
bindings (strong, visible, waived, and not-perceptible) for different
priority levels.

The refactoring includes:

- A `ChildServiceConnectionController` interface defining the contract
  for managing service bindings.
- `LegacyChildServiceConnectionController`, which encapsulates the original
  logic of managing multiple bindings for older Android versions.
- `ChildProcessConnection` is updated to delegate all binding management
  to an instance of the controller.
- State and logic related to binding management (e.g. binding counts,
  binding state, re-binding logic) have been moved from
  `ChildProcessConnection` into the controller implementations.

In this change, the logic of
`ChildProcessConnection.retireBindingsAndBindFallback()` is slightly
changed. It used to abort binding connections if any of connection fails
to bind. However, since the abort is a kind of optimization for the rare
failure case, we can omit it and simply call
`ChildServiceConnectionController.add{level}Binding()` to reconstruct
the binding counters.

Bug: 436110326
Change-Id: Iac0cf6daf917543f18931f555a09663db907f748
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6922420
Commit-Queue: Shin Kawamura <kawasin@google.com>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1514635}


CrOS-Libchrome-Original-Commit: b0df5d550527d8d709333b52cf164fd3c2ac5cd8