Allow StackTrace().ToString() to work when dbghelp not available
In sandboxed processes on Windows dbghelp (and PDB symbols) are not
available so symbolizing a stacktrace never really worked, and crashes
if dbghelp is not already loaded in the process. This does not happen in
many places but in dcheck-canary a dcheck with a stack provided to its
log message is masked (e.g. goto/crash/e2cd3a2bdcddc686) when symbol
initialization fails.

profiling_client would try to initialize the symbolization
infrastructure when a new trace was started. In sandboxed processes this
will crash if dbghelp is not already loaded, so this CL avoids calling
this in official windows builds (tested manually).

This CL allows base::debug::StackTrace to do something useful in these
processes (output a string of addresses) & adds tests.

Bug: 443286263
Change-Id: I6c3f8dd18fbeb40799e61291af8471dadaf840e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6933898
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1515089}


CrOS-Libchrome-Original-Commit: 0bbe87d347dc4af926859752550ee7c9be476e39