[mojo] Reserve space in DictValue/ListValue on unserialization.
DictValue/ListValue unserialization consistently shows up on jank
flame graphs. Reserving space in the underlying containers to avoid
re-allocation will probably not fix the problem entirely, but it's
a net improvement. We make the change behind a Feature to help us
understand whether the steps we take to improve performance in the
field are going in the right direction (the field is affected by
memory contention, which is difficult to reproduce perfectly in the
field).

Bug: 424432184
Change-Id: Ia95a5243798935f19a38218be7fb6d62c129a8ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6900502
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1516031}


CrOS-Libchrome-Original-Commit: 2a9a8d845971a754ffcc30bffef724cabe51e544