[android] Increase priority of network thread when PI locks are enabled
Local testing shows that when priority inheritance (PI) locks are
enabled on Android, the network service thread is frequently queued
behind thread pool worker threads when contending for a PI lock,
regressing startup time. This is due to the Linux kernel enforcing FIFO
ordering on threads of same priority contending on a PI lock.

This commit increases the network thread's priority when PI locks are
enabled to compensate for the shift from an unfair to a fair lock.

Also, update the corresponding synthetic field trial's version suffix.

Bug: 438170467
Change-Id: I39d59b62b03aa02eaeedf6a324c217343069bff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6961497
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Anand Ravi <anandrv@google.com>
Cr-Commit-Position: refs/heads/main@{#1517292}


CrOS-Libchrome-Original-Commit: ccb97eea00a3cba0c76fd3196a5626a1a3198326