Enable PartitionAlloc External Metadata feature at 100%
This feature moves metadata from inline in the gigachage to a different
address. This aims to harden PartitionAlloc's metadata from external
threats (like inside the V8 Sandbox memory corruption).

Previously this feature was being tested at 50/50 but was disabled on
Monday (crrev.com/c/6940460) which was released Sept 16th in Canary
(142.0.7416.0), this has given perf bots a couple days without this
feature. Removing the effect of the 50/50 experiment. We will now enable
this feature at 100% to allow perf bots to determine if there are
noticeable regressions before disabling it again before the Beta branch
cut on Sept 29th.

We will then use this along with the 50/50 beta experiment results to
determine if we should launch this properly in M142/M143.

Bug: 435341877
Change-Id: I8ba12d8be9a311d13d3e3d8ed29cd52cc091c1c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6964115
Reviewed-by: Takashi Sakamoto <tasak@google.com>
Commit-Queue: Takashi Sakamoto <tasak@google.com>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1517797}


CrOS-Libchrome-Original-Commit: 4c10aff2ca6a34fa731715c4ce4c060d45644dc8