Add a THREAD_CHECKER and threading annotations to ThreadLocalNode
This better documents which methods must be called on the thread that
ThreadLocalNode is bound to. It doesn't change any of the existing
semantics of DirectReceiver or ThreadLocalNode.

The `task_runner_` member is no longer needed since the THREAD_CHECKER
replaces it outside the constructor.

R=pasko

Bug: 446199357
Change-Id: I377e1f8af703353e985add054be92532bc57c031
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6964881
Auto-Submit: Joe Mason <joenotcharles@google.com>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1518178}


CrOS-Libchrome-Original-Commit: 2c23f2735176e8958428c9bfb90cbd2df79f07cc