Fix libchrome compile error in MessagePumpGlib
Upreving libchrome on ChromeOS failed with a compile error. libchrome is
built with use_glib=true, which includes message_pump_glib.cc, but
without IS_OZONE defined. This caused a call to WatchFileDescriptor
which was not declared.

This change guards the call in message_pump_glib.cc with
BUILDFLAG(IS_OZONE) and simplifies the matching guard on the declaration
in current_thread.h to resolve the build failure.

Bug: 379190028
Test: Tryjob
Change-Id: I70217eb2e617bbb12b8123e3c9ed488c41eee3c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6956308
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Di Wu <diwux@google.com>
Cr-Commit-Position: refs/heads/main@{#1521713}


CrOS-Libchrome-Original-Commit: 94ee304db4cf71eec2e27d0b3324a9ba10ca644d