[webview] Add experimental Library Prefetch configuration API
Whilst renderer native library preloading is generally a performance
improvement in most scenarios, it can regress the performance of more
important app tasks if it ends up competing for CPU, memory, or disk
resources.

This change adds an experimental API to allow apps to globally
configure whether library prefetching should be performed by the
renderer. It can be changed dynamically throughout the app lifetime to
accommodate different scenarios.

Note that this is not a finalized API, but should be suitable for
performance evaluation purposes. A more polished API would likely not
apply globally, but instead specifically to either a Profile or WebView
instance, or perhaps something like WebViewRenderProcess.

The API is intended to be exposed via AndroidX as
androidx.webkit.WebViewCompat.setRendererLibraryPrefetchMode and
androidx.webkit.WebViewCompat.getRendererLibraryPrefetchMode.

The implementation is given a feature flag, which can be used to turn
the API into a no-op.

A metric is added to provide visibility into the prefetching modes
that apps start renderers with.

Bug: b:428680502
Change-Id: Ia7d9264b5a5aa9c443975b8be4b8bddffd886da9
Enabled-by-default-reason: killswitch
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6966406
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Commit-Queue: Ashley Newson <ashleynewson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1523713}


CrOS-Libchrome-Original-Commit: 166dc701af7d060c54df9f72008e8b4d90399a6d