Add ts typemap for string16
This typemap converters string16 to string and vice versa
automatically. String16 should be completely opaque to ts users after
this.

Unfortunately it is not possible to partial migrations because of how
the bindings are generated. It is a risky change, but we shouldn't have
too many of these mega changes. Url would probably be the next big one.

Fixed: 431824282
Change-Id: I4f5f74ae103b583a50f97ce326fb070f748343c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6987362
Reviewed-by: Michael Cui <mlcui@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1524994}


CrOS-Libchrome-Original-Commit: 12fc15823a476304996ed4ea31d17521808b4d6d