Remove old rust mojo bindings code
This code is being redeveloped from scratch, and will soon start
to be committed here. Get rid of the existing code beforehand to
avoid overlaps and confusion.

We'll end up re-adding some of this, but deleting it all beforehand
ensures we don't have stale code mixed in to the new stuff.

Bug: 444509367
Change-Id: Iae6a4def522b24154befbfe9687869bca9870130
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7016908
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1527031}


CrOS-Libchrome-Original-Commit: cf1413acd174de2f050f17696a53025217b438f2