Add const to boolean operators
We do not want to modify either the left nor right side of the operators.
We also do not want to copy by value the objects for this when the objects
are large and complex.

This CL was uploaded by git cl split.

R=lizeb@chromium.org

Change-Id: I1f376fb5acf55e774e989055a8eafe3f3cbc8fbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7022569
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Commit-Queue: Kaylee Lubick <kjlubick@chromium.org>
Auto-Submit: Kaylee Lubick <kjlubick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1527471}


CrOS-Libchrome-Original-Commit: d1d83ad557707f11afc4e0a2df6535be864588a5