mojo: Ensure Channel::Message::header() is not used with ipcz
A few methods in mojo::core::Message are like this:
```
if (is_legacy_message()) {
  ...
} else {
  // Something calling header()
}
```

These methods are not going to work correctly if they happen to run for
IpczMessage because struct Header is not used with ipcz. I found one of
methods like this being used in ChannelLinux to decide whether fallback
to ChannelPosix is needed:
```
void ChannelLinux::Write(MessagePtr message) {
  if (!shared_mem_writer_ || message->has_handles() || ... ) {
    return ChannelPosix::Write(std::move(message));
  }
  ...
}
```

Currently shared_mem_writer_ is always false, so luckily the incorrect
has_handles() is not used. To prepare for experimentation with the
shared memory writing (and notifying over eventfd) we need to override
num_handles().

To ensure that header() is not accidentally used with ipcz, here we mark
it as NOTREACHED in IpczMessage and in TrivialMessage (the only two
subclasses that can be used with ipcz).

We could mark other methods as NOTREACHED, but that would be verbose.
Instead only change header() and is_legacy_message() - the low level
methods, preventing incorrect use of their transitive callers.

Low-Coverage-Reason: Added a few uncovered NOTREACHED.
Bug: 388531132
Change-Id: I4b0daf3b6ae45c691c66b52f7cf83891d4361836
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7009819
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1527539}


CrOS-Libchrome-Original-Commit: d7db947b673dde5217c7d55e0db1efb455876b75