Add some documentation about sqeuences to raw_ptr.md
The error messages in ASAN dumps (base/memory/raw_ptr_asan_service.cc)
point you to raw_ptr.md. But one of these messages talks about the
requirements for sequences and raw_ptr.md does not. This CL adds some
text to raw_ptr.md to describe the analysis needed in this case. All
this text is technically present with the combination of
raw_ptr_asan_service.cc and raw_ptr.md, but it's nice to be explicit in
the documentation we link from the dump.

Change-Id: I11545df8ba8145e22e29cbedc35d0567d8823aa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7027596
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1528198}


CrOS-Libchrome-Original-Commit: b4cb86f3c825cd7cae6b5d8258d134a5fd3c9c70