Enable strict handle checking in tests
Strict handle checking is already enabled for processes running in an
App Container sandbox, and this will be expanded in the future to all
sandboxed processes.

It is a strong indicator of a bug if an invalid handle is used, so this
CL applies the PROCESS_MITIGATION_STRICT_HANDLE_CHECK_POLICY to all
tests. This is intended to identify any latent issues related to invalid
handle usage that might exist in current code.

Unit tests are covered by adding the policy for the process in
base::TestSuite, and browser tests are covered by enabling the policy
for shell embedder that is used by tests.

Three tests were using invalid handles and now crash - two of them (in
base) are replaced by death tests, and the mojo test is changed to
disable strict handle checking for the child process receiving an
invalid handle.

No changes to production behavior are intended by this CL.

BUG=448025441

Change-Id: I913ab34fb6707ef2447052262f56ee91c04f5864
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7004149
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530574}


CrOS-Libchrome-Original-Commit: 3e755d08e52115e8987f5de6ecd696337c901eb2