[rustmojo] Initial deparser implementation
This adds a minimal serializer for MojomValues. Unlike parsing, it
takes two inputs: the value itself, and the format it will take on the
wire. Annoyingly, this means we have to do lots of extra validation to
ensure the two formats match.

Tested manually, apart from version numbers this outputs the same thing
as the C++ bindings do for my test case.

Bug: 444509367
Change-Id: Ibb0c38498657111068d44be34c06cc78169d2f9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046838
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1532271}


CrOS-Libchrome-Original-Commit: 6d0effa1af4221f31e03fcfa5b0dd72d5cc2d6ac