Pass Bus::Options by value with std::move.
This modifies the dbus::Bus constructor to accept the Options parameter
by value instead of by const-reference. Callers are updated to use
std::move() since Bus::Options is a move-only type.

The intention is to add a new OnceCallback property to Bus::Options in
a followup CL to address the linked bug. This is not possible if
Bus::Options is passed by const-reference.

Bug: b:447203893
Change-Id: I7880c1fd713a53b6c6794c81f389be4a7bf34767
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7056670
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Sun <michaelfsun@google.com>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532789}


CrOS-Libchrome-Original-Commit: 7a20c86873127be0afa5bfb54f2c0cc3a987decb