metrics: clean up use of histogram spare file path.
The pathname for the spare file used by the persistent histogram
allocator was being calculated independently in different places, and
will be needed in a new place in //android_webview for an upcoming
metrics path migration.

Unify the existing places that calculate it, and expose a function to
get it for use by the following WebView change.

The filename and size were also being referred to from Java code in
//chrome in two places, relying on IFTT to keep them in sync for one
place but just a comment for the other; unify these two as well so that
it's only duplicated into Java once.

Change-Id: Id18ff12a5a28e3faa780705364e972b574e4f35e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7055744
Auto-Submit: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533196}


CrOS-Libchrome-Original-Commit: 843cbf981ff1f5442a0fab5f7a3356124238355a