mojo: Experimental way to notify with eventfd on Android
In mojo/core/ipcz_driver/transport.cc, if the base::Feature flag
(kMojoLinuxChannelSharedMem) is enabled, we send a special message from
the broker: "upgrade offer".  The other side will likely respond with
accepting the offer, but may reject the offer (for example, because it
ran out of file descriptors).  Upon 'accepting' the upgrade the peer
will need to send its own upgrade offer. Once both sides accepted the
offer, the channel starts communicating over 2 shared memory regions,
using eventfd to wake up the other side. Importantly, the message may
not fit the shared memory region, so ChannelLinux (the Channel
implementation on Android) falls back to the ChannelPosix sending, which
means that the receiver will need to get this message from the socket.

In the pre-ipcz world the 'upgrade negotiation' and the eventfd/shmem
communication was tested in the field, so I am assuming it is quite
robust.

However, with ipcz there are two problems:
1. The IpczMessage cannot be intercepted invisibly in Channel because it
   lacks the message_type field (used by the eventfd notifier feature to
   distinguish 'control messages' - the ones used for the upgrade
   handshakes)
2. Ipcz is not resilient to messages arriving out of order occasionally,
   which can happen because of the fallback from ChannelLinux to
   ChannelPosix (sending over socket instead of shmem).

To address these problems we extend IpczMessage by |message_type| for
intercepting control messages and |message_count| for reordering
messages. This is extended only on Android to allow faster
experimentation with the wire protocol, avoiding compatibility issues.

Messages arriving too early for their order number are saved and
dispatched later when the gap is filled, using a simple flat_hash_map
mapping the order number to the message. I expect that in practice this
map is going to be very small. No extra locking seems to be needed
because ChannelLinux/ChannelPosix already assume that messages are not
arriving concurrently (only a single thread is using a single epoll_wait
for all the watchers).

This is similar to the "early parcels" change. The early parcels fixed a
part of the problem of wrongly ordered messages, but not fully.

Since the old ChannelLinux shared memory implementation is hooked into
pre-ipcz protocol, make sure to disable this part for old peers: never
send upgrade offers, upon receiving reject them, never 'advertise' the
capability.

Bug: 388531132
Change-Id: I4bace74e24be05ff456b02d5758736206e67a191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7016021
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533524}


CrOS-Libchrome-Original-Commit: 09f667bf5884b80072eda164d56480e1a61d9f4c