absl::Hash support for base::StrongAlias
Implement AbslHashValue() for base::StrongAlias so that it can be used
as a key in Abseil hash tables.

Adjust BUILD.gn and DEPS files to permit AbslHashValue() implementations
to be tested with the absl::VerifyTypeImplementsAbslHashCorrectly()
function. It is helpful to find incorrect hashing implementations.

This CL has the side-effect of moving some types that were using
std::hash specializations to use Abseil's native hash implementation
instead. This is expected to be strictly beneficial.

BUG=452082294

Change-Id: I7c22d7d19146e967d91323b17a878983ff5b7d34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7066374
Reviewed-by: Danil Chapovalov <danilchap@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534537}


CrOS-Libchrome-Original-Commit: 7c96e75b1ecd4d673eefbb4bdd11286d462aa7ee