[rustmojo] Start working on generating code for mojom parsing
This CL lays the groundwork for hooking actual rust types up to the
parser. It primarily does two things:
1. Defines a trait for a rust type that can be (de)serialized to a
mojom message, and implements it for u8 as an example.
2. Sets up the build system so we can start writing a proc macro to
derive the type automatically.

In followup CLs, we'll implement the trait for the rest of the
primitive types, write the proc macro, and then provide a
parsing/deparsing API that accepts and returns objects with that
trait, instead of only Mojom ASTs.

Bug: 444509367
Change-Id: I67908f891c9ee8f196ac7cec120c6fa0f501f75d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7068293
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1538106}


CrOS-Libchrome-Original-Commit: 4cb67716edbfdf7be5ab53499f75f11694261af9