[rustmojo] Implement derive macro for MojoParse trait on structs
This allows us to automatically generate parsing and deparsing code
for "mojom-compatible" structs, i.e. those that match the subset of
mojom we support. Generally, these structs will be generated by the
bindings code.

Although it is possible to put this on user types if they have the
right format, providing some basic type mapping, I don't think it's
going to extend well as we add more mojom features, so we shouldn't
advertise it to users.

Bug: 444509367
Change-Id: I6db965ebff0a9cfe01031888ae9751ef050aa11e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076675
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1538186}


CrOS-Libchrome-Original-Commit: 1089f7c8ac311ae62a867f7a4b524b5b374e4e45