[rustmojo] Rework MojomParse trait to use From and Into
I noticed that the conversion functions to/from MojomValues were
essentially identical to the standard From and Into traits, so
this CL reworks the MojomParse trait to require them, and adjusts
the derivation macro to implement them.

This is mostly a minor structural change, but From and Into are widely
used in rust, so it seems good to be idiomatic.

Bug: 444509367
Change-Id: I89c9a3cc93eb6773ede986c7d167ace9b3806fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076695
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1538188}


CrOS-Libchrome-Original-Commit: 33351d0426b5682cb880065a92f3c44290b8e2fb