[installer] Improve unbuffered file writes
Some devices are reporting 8MiB physical sectors via
GetFileInformationByHandleEx, which seems highly unlikely. This is
leading to unnecessary OOM crashes.

Switch to querying StorageAccessAlignmentProperty from DeviceIoControl
in the hopes that it returns the correct value in more cases. Also,
handle a failure to obtain a reasonable value by assuming 4KiB sectors
rather than failing the operation. If this assumption proves to be
false, a subsequent failed write will be handled property. This should
be visible in failure metrics if it happens in practice.

To be safe, new histrogams are added to report the result of the attempt
to determining the physical sector size and the discovered result.

OBSOLETE_HISTOGRAM[Setup.Install.UnbufferedFileWriter.GetInfo.Error]=Superseded with Setup.Install.UnbufferedFileWriter.DeterminePhysicalSectorSizeResult.

Fixed: 456155453
Change-Id: Id9ac48d2c52818b7952d3922a9093ec4d7e996f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7096263
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538574}


CrOS-Libchrome-Original-Commit: a5bea74177b67f09942494d50d871305a65423db