[Android] Restrict ActivityElements to Stations/CarryOns and to only 1
This is a step to fix ViewElement checking and interactions with
multiwindow tests.

Add a determineActivityElement() to ConditionalState, which will be used
when ViewConditions restrict searching for Views inside an Activity's
subwindows (and dialogs).

Stations already have an associated typed ActivityElement, and
Facilities will use the associated Station's ActivityElement. CarryOns
now have an associated loosely typed ActivityElement.

It is not necessary to declare an ActivityElement if no ViewElement is
declared.

Bug: 414438521
Change-Id: I7b9fab59dac0f421d1389453c88d80e02803f027
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7095890
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538691}


CrOS-Libchrome-Original-Commit: 7ab0826d4ac39b7a51446f0a5430ad9c71027357