[base/message_loop] Remove CHECK that prevents watching the same FD for the same signal in MessagePumpAndroid
When falling back to OS transports for IPC instead of using shared
memory, channel_posix utilizes MessagePumpAndroid's IOWatcher for read
and/or write signals on sockets/FDs.

Imagine we have small socket buffers or the message being written is
very large. In this case, channel_posix will write as much as it can and
then set up a write_watcher_ (one-shot IOWatcher::FdWatch) to be
notified when it can continue writing to the socket.

When the socket is writeable to again, ALooper wakes up and the
ChannelPosix::OnFdWritable callback runs. This callback will try to
finish the previously incomplete write, however this socket write could
also return EAGAIN meaning channel_posix will once again try to set up a
write_watcher_ with the same FD, hitting this CHECK failure. IOWatcher
cleans up its resources for a one-shot watch only after the callback
finishes running, but the first callback hasn't finished yet.

Async I/O watching for MessagePumpEpoll handles this behaviour. In this
case, we try to mimic what happens in MessagePumpEpoll as ALooper is
backed by epoll. When adding an existing FD with ALooper_addFd, the
looper will do an epoll_ctl with EPOLL_CTL_MOD so only the event mask is
updated.

We detach the pre-existing FdWatch so it is no longer responsible for
controlling the lifetime of the active FD watch.

Bug: 442028777
Change-Id: I0d44368916a9d7a416cfbf857a01184b0b6ee121
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7028812
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Nafis Abedin <nafisabedin@google.com>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538727}


CrOS-Libchrome-Original-Commit: 68ed063b566004d35285fa54cb765d1964e221fb