[OOPPD] Fix "Send in Mail" from the system print dialog
For OOPPD on macOS, the print dialog in the browser process captures all
the print settings, and the Print Backend Service applies them. For
"Send in Mail", the destination URL is a file:// URL for a symlink. When
the Print Backend Service tries to apply the URL, macOS does not like
the symlink. Fix this issue by resolving the symlink in the browser
process when capturing the print settings.

Along the way, use `base::apple::ScopedCFTypeRef<T>` more and avoid raw
`T` usage.

Bug: 456225707
Change-Id: Ifb1921732644d2a25e6e75f62ca19886a0d523ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7104899
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538905}


CrOS-Libchrome-Original-Commit: 6f45f0df3c4a52d37df796afbdd49d02cd62568f