[rustmojo] Begin working on actual unit tests for the mojom stack
We've finally got enough implemented to move from our temporary "just do
something for coverage" tests to a more structured format. This CL
starts the process by adding a bunch of tests for the bindings generator.
It ensures that the generator outputs the expected rust type for each
mojom type, and that the MojomParse macro derives the corresponding
ASTs correctly.

The next step will be to test the actual parser/deparser to/from a
binary format. Since we've covered the bindings generator here, we can
re-use it for those tests, and write them in terms of rust types instead
of ASTs.

Bug: 444509367
Change-Id: I8617d11bf255b52ed53db95381652c7323dc94f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7103041
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1540788}


CrOS-Libchrome-Original-Commit: 3963977daa83811c54a2ca10d286f2b0c1373e1c