[Android] MultiWindowAppMenuTest with RobustWindowManagementExp. on+off
Create variants of MultiWindowAppMenuTest with the
RobustWindowManagementExperimental flag enabled and disabled. The
behavior is quite different, so we should cover both cases to some
extent.

Not using parameterized tests so that these can be batched into
two batches, one for each flag value, which requires the
@EnableFeatures/@DisableFeatures annotations.

Bug: 360187206
Change-Id: Icbab2ed3380590c5490fe8bc086d372996930036
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7115218
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Cr-Commit-Position: refs/heads/main@{#1540841}


CrOS-Libchrome-Original-Commit: b0bdd2355227a845bd6fb1ef9cf86074a4c2bafe