mojo: client impersonation/max clients options in NamedPlatformChannel
These are feature changes for Windows clients/servers.

`allow_impersonation`: allows a client to specify that it will allow the
server to impersonate the client. This is to allow for a high privilege
server to impersonate a client. For instance, the server may want to
check if the client is an administrator before allowing certain calls,
or to get client specific settings.

`max_clients`: this setting allows a server to explicitly specify the
maximum number of clients that can connect to it at any given time. This
setting allows `enforce_uniqueness` to only mean enforcing the
uniqueness of the named server, instead of also meaning how many clients
can connect to the server like it currently is overloaded to do.

Fixed: 458202080
Change-Id: I04b64ea482e5219aca37f4a46e04169cff14f25d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7127532
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: S Ganesh <ganesh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541546}


CrOS-Libchrome-Original-Commit: 66a5c384a01aa50aa26b79b8443de70f109cdc8b