libchrome: Prepare for removing MockObjectProxy's Do* methods
Update patch backward-compatibility-3300-Revert-Remove-Do-Methods-in-*
such that it creates mock functions for the "proper" methods, so that we
can migrate client packages.

To not break the current code, the MockObjectProxy sets default handlers
for all mock methods that call the Do* functions. This doesn't work for
a test in the trunks package (due to StrictMock), so make this CL
co-dependent with a change to trunks.

BUG=b:453531230
TEST=cq

Cq-Depend: chromium:7144243
Change-Id: I1d09b93cfd477e65aae3c161a13c0f35272da40a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/libchrome/+/7144519
Tested-by: Georg Neis <neis@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>
Reviewed-by: Di Wu <diwux@google.com>