Move PasswordManagerBlocklistPolicy to //components/password_manager
1. Today, we encourage developer to put their policy implementation
together with the feature code, so move the implementation of
`PasswordManagerBlocklistPolicy` to the correct directory.

And the pref `kPasswordManagerBlocklist` is also moved accordingly.

2. This policy uses URLBlocklist to match URLs, so a corresponding doc
has been added to the policy description to guide users to [1] for more
information.

[1] https://support.google.com/chrome/a?p=url_blocklist_filter_format

3. password_manager_blocklist_policy_browsertest.cc verify nothing
but pref. However, the policy -> pref mapping has already covered by
PasswordManagerBlocklist.json and there is no need to create a separate
browser test for the same thing.

So remove the password_manager_blocklist_policy_browsertest.cc test.

4. Update BUILD.gn and DEPS files accordingly.

Bug: 408018852
Change-Id: Ie74a5e21bb4f3c72231fa735f52c4fc2c5094426
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7155470
Commit-Queue: Ho Cheung <hocheung@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547046}


CrOS-Libchrome-Original-Commit: 08b0ed5ebca7d7fc0a6c37dfd3fdb6628ee80504