GpuPersistentCache: Make reads wait max 20ms for writes to finish
Currently a PersistentCache instance cannot do reads & writes in
parallel. Hence if a write takes a long time to finish, any reads happen
during this duration will be blocked.

To avoid long reads' waiting time, we will use external lock & condvar
to wait for a maximum of 20ms. If timeout, we will treat the read as
cache miss. This is to allow the calling thread to continue. This is
a trade off, if the wait is too long, it's better to let the calling
thread re-do the shader compilation.

Bug: 399642827, 439845637
Change-Id: I8c6adeb7e82f0ad4e552ce90925356f71b45d4ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7130101
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547343}


CrOS-Libchrome-Original-Commit: 076b8b99b9729d4266e43df22fc96d06eb34421a