Add autoconf to install-build-deps.py
This is required by the GN linux bot. That bot is shared with various
other projects, such as pdfium, boringssl, etc (i.e. the bot is in the
flex pool). So the easiest way out is to install the package for all
bots. See more context in [1] and [2].

[1] https://crrev.com/i/infra/infra_internal/+/8764696/comments/e8449541_d2b8abce
[2] https://crrev.com/i/infra/infra_internal/+/8764696/comment/bc782619_b2313a2f/

Bug: 459541007
Change-Id: Ie0def35214bece4543c895666b6ad8f3356da95e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172045
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547555}


CrOS-Libchrome-Original-Commit: e640d3e88551b6a7dba3bc31458931358861d7f0