[rustmojo] Add support for mojom enums
This CL allows us to handle mojom files that define enums. We don't
currently support extensible enums, or those which are nested inside
a struct or interface.

The key property we need from enums is the ability to convert back
and forth between their 'enum' representation and the u32 discriminant.
We encapsulate this ability in the PrimitiveEnum trait, which can be
derived automatically for Rust enums whose variants don't carry any
data. In the future, we could get similar functionality by importing
the num_enum crate, though we'd likely still need a wrapper trait.

We include blanket implementations for any type that implements
PrimitiveEnum, including for MojomParse, so you don't need to derive
those traits directly.

Touches a bunch of files, but most of the changes are small or tests.

Bug: 444509367
Change-Id: I6e9579435d575992ad2f4513ae2c451ab189548e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7138878
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1547888}


CrOS-Libchrome-Original-Commit: c0291bebd7404fb17bc8ed36c5e85ebe0a875172