[MemoryPressureListener] Prepare for sending no pressure notifications
The current API contract is that OnMemoryPressure is not called when returning to the MEMORY_PRESSURE_LEVEL_NONE state.

This will change in the future, so current implementations must allow
it, even if they don't do anything with it.

Bug: 462288601
Change-Id: I50149997d113be86456f221482b41235f4ddfe2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7168848
Owners-Override: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548666}


CrOS-Libchrome-Original-Commit: aef3061376f1a0a6ec30d28eafb842ab712e697c