mojo: Fix transport leak in MojoSendInvitation failure path
In `Invitation::Send`, an ipcz driver handle for the transport is
created before validating the consistency of the provided attachments.

If the attachment validation fails the function returns
`MOJO_RESULT_FAILED_PRECONDITION` immediately. Previously, this path did
not clean up the `transport` handle created earlier, resulting in a leak
of the underlying platform handle.

This change ensures the transport handle is closed if the function
returns early.

Bug: 449764077
Change-Id: Ibfa2b5bb9586cc71bc64f443ed5a6a0b328fbfe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7206262
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1552328}


CrOS-Libchrome-Original-Commit: afe7683ca6a7c6368f179a7ae16b18c454936a24