[rustmojo] Add simple data pipe producer + consumer test.
A follow-on CL creates support for two-phase reads/writes, but it was
rather large to do it all in one CL, so this CL gets the most basic form
of functionality in place.

Additionally drops a placeholder trap.rs file, as implementation of
trap/wait are necessary for the two-phase reads/writes (also in a
follow-on CL).

Bug: 444509367
Change-Id: I6e7d344dd8b5600a463bdeb0decb0e41bb67ae1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172099
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Reviewed-by: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1552485}


CrOS-Libchrome-Original-Commit: 13a341bc432eccd963580fcac9077149c400ad0b