[rustmojo] Separate ordinals from wire types
This is a refactor in anticipation of supporting arrays. Doing so
requires us to generalize our representation somewhat. Previously we
only considered wire types in the context of a struct body, and
attached an ordinal to each one. However, in other bodies, the ordinal
was meaningless, and having it attached to the wire type introduced
lifetime issues when we wanted to create wire types on the fly.

This separates ordinals into their own type, which is either a bitfield
with many ordinals, or a single value with one ordinal. This also
addresses some of the weirdness that bitfields introduced, since they
were the only value with multiple ordinals. Now, bitfields only
exist in the context of a structured body, and we've introduced a Bool
type for booleans in other contexts.

As another refactoring prep, we change the `fields` argument when
parsing/deparsing bodies to be an iterator instead of a slice. This
is because iterators give us more flexibility with the layout of the
source data, while slices require contiguous chunks.

Bug: 444509367
Change-Id: I17953a13bdcafa9c446ab50a91241610c6a36710
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7188792
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1554667}


CrOS-Libchrome-Original-Commit: e01c18d8a90cb57a952938e7da109490c8282f76