[rustmojo] Add support for mojom strings
This CL allows our parser to serialize and deserialize strings to/from
mojom. The main difficulty here was deciding what representation strings
should have in rust. Mojom wants its strings to be UTF8 encoded, but
they aren't, so the String type is not suitable. The OsString type is
nearly suitable, but it uses an unspecified encoding and so can't be
safely sent across the wire (e.g. to C++ code).

Ultimately, we settle on using a wrapper type `MojomString` which just
provides a thin shell around a byte array. We provide several conversion
functions, but users will be forced to handle non-UTF8-ness explicitly
when retrieving parsed values.

Bug: 444509367
Change-Id: I484ae29574f2e3117f47b8fec71b398d7864205a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7217508
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1554749}


CrOS-Libchrome-Original-Commit: 34ecdbb7ce08e5d85f615afec5d2cf478211bca4