[rustmojo] Add support for nullable values
This CL adds support for the final mojom type (besides handles):
nullables. Mostly, this involves putting checks after parsing/before
deparsing, to check if the value is nullable and wrap/extract it as
appropriate.

Nullables in mojom have two representations, depending on whether they
are primitives or not. Primitives are encoding using a tag bit, which
is packed as a completely separate boolean from the actual value. Non-
primitive values just use 0 to represent a null value.

There are also a few tweaks to older code that I ran into while working
on this.

Because this CL is very large, the bindings generation and testing code
is in a followup CL. Sadly, I wasn't able to pare this one down any
further (since changes to the AST cascade everywhere).

Bug: 444509367
Change-Id: I923f957f347c2f7fa7d91192f3ceb29c6e865cb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7233331
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1554955}


CrOS-Libchrome-Original-Commit: 443101268edd9f45450318438556ee8bb2a41348