Improve testing async MemoryPressureListeners and MemoryConsumers
A challenge with testing async MemoryPressureListeners and
MemoryConsumers is that registration is also async, and the
notification itself is also async. But a test usually need to wait
until the notification is received to proceed.

This CL modifies SimulatePressureNotificationAsync to allow waiting on
the notification, and add equivalent functionality to
TestMemoryConsumerRegistry.

Bug: 441951621
Change-Id: I687e93cdc277482f3ec8de6c9f283f58279f5b7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7229564
Owners-Override: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1554986}


CrOS-Libchrome-Original-Commit: 1294cc6730f8b5956c8e81958a41d389fc046412