Using DeleteSoon for async registrations in memory pressure observers
I initially didn't use DeleteSoon because I wanted to ensure that the
unregistrations would always be executed, even in the case where the
task queue is cleared without running.

This approach isn't guaranteed to work because the task can also be
deleted if it is rejected by the PostTask() call, which would result in
the unregistration running in a non-thread-safe manner.

An issue with DeleteSoon, is that the object is leaked if the task never
ends up being run. While this is the behavior that we want in production
code, this cause the leak sanitizer to complain in tests where the
TaskEnvironment is not correctly flushed.

Bug: 406578344
Change-Id: I7c43e896e1b985aba045dcb5cf52b4d626d74e30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7252090
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558155}


CrOS-Libchrome-Original-Commit: 003ead1d49ab9485cf30d4d1b013cb07510fd221