Reland #2 "Add a feature param to initialize feature list early for Android"
This CL adds a new feature param called InitFeatureListEarly.
When it is enabled, we will initialize feature list early, immediately
after loading the native library. In order to do that, we will need to
initialize a number of things early, such as creating the thread pool
and creating the variations ids provider.

In this version, we check whether the feature list has been initialized
early via a boolean flag, and then decide whether content main needs to
initialize feature list. This is different from before where we use
the feature flag and param's value to determine whether content main
needs to initialize feature list.

Also in this version, we have decided to only initialize the feature
list early in one test; all other tests will exercise the previous
behaviour. This will need to be the case, because we are planning to
support first run flow by checking whether there is a Finch seed.
If there is a Finch seed, we initialize feature list early.
But in tests, there is never a Finch seed, so we are never going to
initialize feature list early in tests. This CL introduces a test
called InitFeatureListEarlyTest that specifically tests the early
initialization code path.

A doc with more explanations of what is going on:
https://docs.google.com/document/d/1UNx_fgFjGcp9bjQEMfP_K4Ly1f7-JCO4SNHJN4oBF20

[Gardeners]: Please don't revert this CL for simple test failures,
since this CL is a large-scale change and a few simple test failures
are expected. If this CL does cause a few test failures, please just
disable the tests and assign a bug to me. Thanks a lot for your help!

Bug: 455890560
Change-Id: Ia604acc54148288550cb79d3e5eee7d866550ced
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7265587
Commit-Queue: Martin Kong <martinkong@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Owners-Override: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559662}


CrOS-Libchrome-Original-Commit: 184bbf17a36a27baecb29e9463bf2bba52995f1a