Explicitly instantiate the MemoryPressureListenerRegistry
A future change will add state to the MemoryPressureListenerRegistry,
allowing the notification of the current memory pressure level upon
registration of a listener.

If the registry is a static global that is never destroyed, then that
state leaks to other unit tests.

This CL changes the registry to require explicit instantiation. This
way, unit tests that actually need a registry can instantiate it
themselves, and the state does not leak to other unit tests.

Bug: 462288601
Change-Id: I37df95d21cadd3ef2ccbb6a51716481920e48cd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7243590
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Owners-Override: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560019}


CrOS-Libchrome-Original-Commit: 83f40857ad7750ffec0e4e176a0bb73b5c0f2682