[LRUCache] Support a maximum size of zero.
This allows a user of the cache to temporarily disable the cache, which
will be useful when chrome is under memory pressure.

To support this, the max size was changed to use std::optional, to
support the existing NO_AUTO_EVICT mode, which used zero as a sentinel
value. Now, std::nullopt is used for NO_AUTO_EVICT.

One consequence of this change is that now Put() can fail, iff the
maximum size is zero.

I manually audited all callers to ensure none was using the literal zero
to engage the NO_AUTO_EVICT mode, as this now means a very different
thing.

Additionally, it seems like the BitmapFetcherService class intentionally
wanted to disable its cache on Android by setting the max size to zero,
but unintentionally enabled the NO_AUTO_EVICT mode. Now the LRUCache
behaves as expected, but the associated test had to be disabled.

Bug: 465068849
Change-Id: Id3a6db4029f849bf0a67cca784b7a7c2534b34d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7228584
Reviewed-by: Matt Mueller <mattm@chromium.org>
Owners-Override: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560523}


CrOS-Libchrome-Original-Commit: 119215c7b1495a1149f1433bc3948da57dab6acf